Combining telemetry and control MQTT clients
