+++ src/net/mqtt_control.rs
@@ -16,6 +16,7 @@ use core::fmt::Write;
 use embedded_hal::blocking::delay::DelayUs;
 use heapless::String;
 use minimq::{Property, QoS};
+use serde::Serialize;
 
 /// Specifies an action to take on a channel.
 #[derive(serde::Deserialize, Debug)]
@@ -105,16 +106,17 @@ impl Response {
 }
 
 /// Represents a means of handling MQTT-based control interface.
-pub struct ControlState {
-    mqtt: minimq::Minimq<NetworkStackProxy, SystemTimer, 256, 1>,
+pub struct ControlClient {
+    mqtt: minimq::Minimq<NetworkStackProxy, SystemTimer, 512, 1>,
     subscribed: bool,
     control_topic: String<64>,
+    telemetry_prefix: String<128>,
     default_response_topic: String<64>,
     delay: AsmDelay,
 }
 
-impl ControlState {
-    /// Construct the MQTT control state manager.
+impl ControlClient {
+    /// Construct the MQTT control manager.
     pub fn new<'a>(
         broker: minimq::embedded_nal::IpAddr,
         stack: super::NetworkStackProxy,
@@ -127,6 +129,9 @@ impl ControlState {
         let mut control_topic: String<64> = String::new();
         write!(&mut control_topic, "dt/sinara/booster/{}/control", id).unwrap();
 
+        let mut telemetry_prefix: String<128> = String::new();
+        write!(&mut telemetry_prefix, "dt/sinara/booster/{}/telemetry", id).unwrap();
+
         let mut default_response_topic: String<64> = String::new();
         write!(&mut default_response_topic, "dt/sinara/booster/{}/log", id).unwrap();
 
@@ -134,9 +139,34 @@ impl ControlState {
             mqtt: minimq::Minimq::new(broker, &client_id, stack, SystemTimer::default()).unwrap(),
             subscribed: false,
             control_topic,
+            telemetry_prefix,
             default_response_topic,
             delay,
         }
+    }
+
+    /// Publish telemetry for a specific channel.
+    ///
+    /// # Args
+    /// * `channel` - The channel that telemetry is being reported for.
+    /// * `telemetry` - The associated telemetry of the channel to report.
+    pub fn report_telemetry(&mut self, channel: Channel, telemetry: &impl Serialize) {
+        let mut topic: String<64> = String::new();
+        write!(&mut topic, "{}/ch{}", self.telemetry_prefix, channel as u8).unwrap();
+
+        let message: String<1024> = serde_json_core::to_string(telemetry).unwrap();
+
+        // All telemtry is published in a best-effort manner.
+        self.mqtt
+            .client
+            .publish(
+                topic.as_str(),
+                &message.into_bytes(),
+                minimq::QoS::AtMostOnce,
+                minimq::Retain::NotRetained,
+                &[],
+            )
+            .unwrap();
     }
 
     /// Handle the MQTT-based control interface.
